Rename old implementation.
